Fixed IntakeOutTake controller to work for manual as requested by drivers. Detailed state machine transitions written up in getNextState() and state transitions with delays for slides and flip to complete added. Use FLIP_WAIT_TIME_LIMT and SLIDE_WAIT_TIME_LIMIT to fine tune wait times.
Manual programmed to
"A" to prepare for outtake
Joystick to raise or lower slides when in outtake mode
"B" to return to Intake

TODO: Color Sensors still. This version of code waits for 2 seconds before tucking wrist when in intake position (note - only after init) later it always tucks - fix the temporary timer reset if required. Once replaced with color sensing, this work properly.