Camera calibratio program added.
Fixed sine and cosine issue when calculating X & Y from bearing and Range. X = Range * cosine (Bearing); Y = Range * sine(Bearing)

Fixed Sandbox to enable AprilTag position output. Put the robot in front of a tag (3 for kemera and 4 for eyeball) and run Sandbox in April mode. This will start printing the current position of the robot. Measure and make sure this adds up. When you position the camera at Bearing 0, its X distance should be the same as what is measured.